Normalized format of generated social embed titles
- Facebook, Twitter, and Instagram embed now generate a title from
oEmbed data, simlilar to: 'Author Name - January 15, 2016 - First ten
words of post...'. This format contains all information that could be
needed to search for a post and provides a solid starting point for
users when coming up with a title. Titles are still editable and a
generated title will not be used if a title was previously entered.